windows: Fix printf format warning

+++ src/brickd/client.h
@@ -1,6 +1,6 @@
 /*
  * brickd
- * Copyright (C) 2012-2014 Matthias Bolte <matthias@tinkerforge.com>
+ * Copyright (C) 2012-2014, 2016 Matthias Bolte <matthias@tinkerforge.com>
  * Copyright (C) 2014 Olaf Lüke <olaf@tinkerforge.com>
  *
  * client.h: Client specific functions
@@ -81,7 +81,7 @@ struct _Client {
 
 #define CLIENT_SIGNATURE_FORMAT "N: %s, T: %s, H: %d, A: %s"
 #define client_expand_signature(client) (client)->name, (client)->io->type, \
-	(client)->io->handle, client_get_authentication_state_name((client)->authentication_state)
+	(int)(client)->io->handle, client_get_authentication_state_name((client)->authentication_state)
 
 void pending_request_remove_and_free(PendingRequest *pending_request);
